scripts/requirements.txt
tuspy>=0.2.2
PyVimeo
PyYaml
